Refactor CI workflows and project structure: update pre-commit hook, reorganize App component, and implement new CI configuration
