--- .github/workflows/test.yml
@@ -1,36 +1,0 @@
-name: Expo Tests
-
-# Run Jest test on push and pull request events on main and dev branches.
-on:
-  push:
-  pull_request:
-    branches: [main, dev]
-
-jobs:
-  test:
-    name: Run Tests
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: '20'
-
-      - name: Cache npm dependencies
-        uses: actions/cache@v4
-        with:
-          path: |
-            ~/.npm
-            node_modules
-          key: ${{ runner.os }}-node-${{ hashFiles('**/package-lock.json') }}
-          
-      - name: Install dependencies
-        run: npm install
-
-      - name: Run tests
-        run: npm run testFinal
-        env:
-          CI: true
