+++ .husky/pre-commit
@@ -1,1 +1,1 @@
-npx lint-staged
+npm run lint:fix && npm run format && git add .
