+++ __tests__/App-test.tsx
@@ -1,5 +1,5 @@
 import { render } from "@testing-library/react-native";
-import App from "../App";
+import App from "../src/App";
 
 jest.mock("../global.css", () => "");
 describe("<App />", () => {
