+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { StatusBar } from "expo-status-bar";
 import { StyleSheet, Text, View } from "react-native";
-import "./global.css";
+import "../global.css";
 
 export default function App() {
   return (
